LRAC-12450 analytics-settings-rest Move constants to api module
