+++ modules/apps/analytics/analytics-settings-rest-api/src/main/resources/com/liferay/analytics/settings/rest/constants/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
