+++ modules/apps/analytics/analytics-settings-rest-impl/src/main/java/com/liferay/analytics/settings/rest/internal/resource/v1_0/FieldResourceImpl.java
@@ -15,10 +15,10 @@
 package com.liferay.analytics.settings.rest.internal.resource.v1_0;
 
 import com.liferay.analytics.settings.configuration.AnalyticsConfiguration;
+import com.liferay.analytics.settings.rest.constants.FieldAccountConstants;
+import com.liferay.analytics.settings.rest.constants.FieldPeopleConstants;
+import com.liferay.analytics.settings.rest.constants.FieldProductConstants;
 import com.liferay.analytics.settings.rest.dto.v1_0.Field;
-import com.liferay.analytics.settings.rest.internal.constants.FieldAccountConstants;
-import com.liferay.analytics.settings.rest.internal.constants.FieldPeopleConstants;
-import com.liferay.analytics.settings.rest.internal.constants.FieldProductConstants;
 import com.liferay.analytics.settings.rest.manager.AnalyticsSettingsManager;
 import com.liferay.analytics.settings.rest.resource.v1_0.FieldResource;
 import com.liferay.portal.kernel.log.Log;
